added exceptions to Passenger and moved logic to controllers

+++ src/main/java/com/example/backendairport/service/PassengerService.java
@@ -1,10 +1,13 @@
 package com.example.backendairport.service;
+import com.example.backendairport.exception.ResourceNotFound;
 import com.example.backendairport.model.Airplane;
 import com.example.backendairport.model.Passenger;
 import com.example.backendairport.repository.AirplaneRepository;
 import com.example.backendairport.repository.PassengerRepository;
 import com.example.backendairport.controller.request.PassengerCreationRequest;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
+import org.springframework.web.server.ResponseStatusException;
 import java.util.List;
 
 
@@ -26,18 +29,23 @@ public class PassengerService{
     }
 
     public Passenger findById(Long aLong) {
-        return passengerRepository.findById(aLong).get();
+        return passengerRepository.findById(aLong).orElseThrow(() -> new ResourceNotFound("Passenger doesn't exist"));
     }
 
     public void deleteById(Long aLong) {
+        if (!passengerRepository.existsById(aLong)){
+            throw new ResourceNotFound("Passenger doesn't exist");
+        }
         passengerRepository.deleteById(aLong);
     }
 
-    public Passenger save(Passenger newPassenger)  {
-        return passengerRepository.save(newPassenger);
-    }
-
     public Passenger updatePassengerDetails(Long passengerId, PassengerCreationRequest passengerCreationRequest) {
+        if (!passengerRepository.existsById(passengerId)){
+            throw new ResourceNotFound("Passenger doesn't exist");
+        }
+        if (!airplaneRepository.existsById(passengerCreationRequest.getAirplaneId())){
+            throw new ResourceNotFound("You need to add an Existing Plane");
+        }
         Passenger passengerToBeUpdated = passengerRepository.findById(passengerId).get();
         Airplane airplaneToBeAdded = airplaneRepository.findById(passengerCreationRequest.getAirplaneId()).get();
         passengerToBeUpdated.setName(passengerCreationRequest.getName());
@@ -46,5 +54,21 @@ public class PassengerService{
         passengerToBeUpdated.setTicketPrice(passengerCreationRequest.getTicketPrice());
         passengerToBeUpdated.setAirplane(airplaneToBeAdded);
         return passengerRepository.save(passengerToBeUpdated);
+    }
+
+    public Passenger save(Long airplaneId, String name, int passportNumber, int ticketNumber, int ticketPrice) {
+        if (!airplaneRepository.existsById(airplaneId)){
+            throw new ResourceNotFound("You need to add an Existing Plane");
+        }
+        Airplane airplaneToBeAssociated = airplaneRepository.getById(airplaneId);
+        Passenger newPassenger = Passenger
+                .builder()
+                .name(name)
+                .passportNumber(passportNumber)
+                .ticketNumber(ticketNumber)
+                .ticketPrice(ticketPrice)
+                .airplane(airplaneToBeAssociated)
+                .build();
+         return passengerRepository.save(newPassenger);
     }
 }
